Macro systems testing, encoder scaling for less pulses, macro analysis reveals that PIC data is dropping out at certain points and requires a reset of the PIC processor to recover... look at timing of the packets on the PIC, at the time of the packet dropping, the PIC still gets data to the control box, but data is failing to get to the Nav.
Macro system on the NAV needs changed to not use the average of the left and the right, as this allows the control system to get to a state where the left and the right distances average to the set point but since neither of them are actually near the setpoint the control system stalls. This is due to the fact that the commanded speed is 0 due to the average showing a small difference to the set point while the left and the right are both "not finished"

+++ Control/Variables.h
@@ -27,8 +27,8 @@ float milliamphours;
 float volts = 2600;
 float milliamps = 200, watts;
 //Encoder Variables
-float encoderR=9999;
-float encoderL=9999;
+long encoderR=9999;
+long encoderL=9999;
 unsigned int encoderSpeedR=9999, encoderSpeedL=9999;
 //Motor Values
 int motorL=85;
